0221-Maximal-Square/Solution_test.go
package leetcode

import "testing"

func Test_maximalSquare(t *testing.T) {
	type args struct {
		matrix [][]byte
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		{
			"test-1",
			args{nil},
			0,
		},
		{
			"test-2",
			args{[][]byte{
				{'0'},
			}},
			0,
		},
		{
			"test-3",
			args{[][]byte{
				{'1'},
			}},
			1,
		},
		{
			"test-4",
			args{[][]byte{
				{'0', '1'},
				{'1', '0'},
			}},
			1,
		},
		{
			"test-5",
			args{[][]byte{
				{'1', '0', '1', '0', '0'},
				{'1', '0', '1', '1', '1'},
				{'1', '1', '1', '1', '1'},
				{'1', '0', '0', '1', '0'},
			}},
			4,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := maximalSquare(tt.args.matrix); got != tt.want {
				t.Errorf("maximalSquare() = %v, want %v", got, tt.want)
			}
		})
	}
}
